Remove BaseMessage by pulling up to Message
