+++ core/src/main/java/org/bitcoinj/core/FilteredBlock.java
@@ -37,7 +37,7 @@ import java.util.Objects;
  * 
  * <p>Instances of this class are not safe for use by multiple threads.</p>
  */
-public class FilteredBlock implements BaseMessage {
+public class FilteredBlock implements Message {
     private final Block header;
 
     private final PartialMerkleTree merkleTree;
