+++ core/src/main/java/org/bitcoinj/core/Pong.java
@@ -28,7 +28,7 @@ import java.nio.ByteBuffer;
  * <p>
  * Instances of this class are immutable.
  */
-public class Pong implements BaseMessage {
+public class Pong implements Message {
     private final long nonce;
 
     /**
