+++ core/src/main/java/org/bitcoinj/core/VersionMessage.java
@@ -48,7 +48,7 @@ import static org.bitcoinj.base.internal.Preconditions.check;
  * 
  * <p>Instances of this class are not safe for use by multiple threads.</p>
  */
-public class VersionMessage implements BaseMessage {
+public class VersionMessage implements Message {
 
     /** The version of this library release, as a string. */
     public static final String BITCOINJ_VERSION = "0.18-SNAPSHOT";
