+++ core/src/test/java/org/bitcoinj/core/BitcoinSerializerTest.java
@@ -236,7 +236,7 @@ public class BitcoinSerializerTest {
     public void testSerializeUnknownMessage() throws Exception {
         MessageSerializer serializer = MAINNET.getDefaultSerializer();
 
-        Message unknownMessage = new BaseMessage() {
+        Message unknownMessage = new Message() {
             @Override
             public void bitcoinSerializeToStream(OutputStream stream) {}
         };
